Drop UtteranceRecogzedRule
