+++ libraries/Microsoft.Bot.Builder.Dialogs.Rules/Schemas/Dialogs/Microsoft.AdaptiveDialog.schema
@@ -25,6 +25,5 @@
         }
     },
     "required": [
-        "rules"
     ]
 }
